refactor(exception): 예외 처리 관련 변경
- BusinessException이 기능 요구사항에 맞게 IllegalArgumentException을 발생시키도록
- Lotto가 커스텀 예외를 발생하도록
