+++ src/main/java/lotto/common/BusinessException.java
@@ -1,6 +1,6 @@
 package lotto.common;
 
-public class BusinessException extends RuntimeException {
+public class BusinessException extends IllegalArgumentException {
     protected BusinessException(String message) {
         super(message);
     }
