+++ src/main/java/lotto/domain/shared/exception/InvalidLottoNumbersException.java
@@ -1,7 +1,9 @@
 package lotto.domain.shared.exception;
 
-class InvalidLottoNumbersException extends RuntimeException {
-    public InvalidLottoNumbersException(String message) {
-        super(message);
+import lotto.common.BusinessException;
+
+public class InvalidLottoNumbersException extends BusinessException {
+    public InvalidLottoNumbersException(ErrorType error) {
+        super(error.description());
     }
 }
