+++ src/main/java/lotto/domain/shared/vo/Lotto.java
@@ -4,6 +4,8 @@ import java.util.Collection;
 import java.util.List;
 import lotto.domain.lottoGame.PickLottoNumbersStrategy;
 import lotto.domain.shared.constant.LottoRules;
+import lotto.domain.shared.exception.ErrorType;
+import lotto.domain.shared.exception.InvalidLottoNumbersException;
 
 public class Lotto {
 
@@ -26,13 +28,13 @@ public class Lotto {
 
     private void requireValidNumberCount(List<Integer> numbers) {
         if (numbers.size() != LottoRules.LOTTO_BALL_COUNT.value()) {
-            throw new IllegalArgumentException("[ERROR] 로또 번호는 6개여야 합니다.");
+            throw new InvalidLottoNumbersException(ErrorType.INVALID_NUMBER_COUNT);
         }
     }
 
     private void requireUnique(List<Integer> numbers) {
         if(hasDuplicates(numbers)) {
-            throw new IllegalArgumentException("[Error]유니크해야해");
+            throw new InvalidLottoNumbersException(ErrorType.DUPLICATES_EXIST);
         }
     }
